requirements.txt
# Requirements for the project
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
EOL